fix(cell): simplify universe deleter and move integration tests
