+++ tests/test_universe.py
@@ -565,52 +565,6 @@ class TestFill:
             else:
                 assert old_val == new_val
 
-
-def test_cell_universe_nullify_with_problem():
-    import montepy
-
-    p = montepy.MCNP_Problem("tests/inputs/test.imcnp")
-    p.parse_input()
-    c = list(p.cells)[0]
-
-    assert c.universe.number != 0
-    c.universe = None
-    assert c.universe.number == 0
-
-    c = list(p.cells)[1]
-    assert c.universe.number == 0
-    del c.universe
-    assert c.universe.number == 0
-
-
-def test_cell_universe_nullify_adds_zero_universe():
-    import montepy
-
-    p = montepy.MCNP_Problem("out2.imcnp")
-    c = montepy.Cell()
-    c.number = 1
-    p.cells.append(c)
-
-    assert 0 not in p.universes.numbers
-    c.universe = None
-    assert 0 in p.universes.numbers
-    assert c.universe.number == 0
-
-
-def test_cell_universe_delete_adds_zero_universe():
-    import montepy
-
-    p = montepy.MCNP_Problem("out2.imcnp")
-    c = montepy.Cell()
-    c.number = 1
-    p.cells.append(c)
-
-    assert 0 not in p.universes.numbers
-    del c.universe
-    assert 0 in p.universes.numbers
-    assert c.universe.number == 0
-
-
 def test_cell_universe_nullify_no_problem():
     import montepy
